+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.9.0] - 2024-01-13
 
 - Add support for reading config from a local JSON datafile [#42]
 - Add support for setting and using default contexts [#43]